src/requirements.txt
flet
arrow 
python-dotenv
requests==2.28.1
